[bucket] add "friends" concept
When you preform positive interactions with bucket, it will increase your
"friendly level", and it will decrease with negative interactions.

Bucket will also greet people on join now, but only if it had seen them before,
and it will be more likely to greet people who have a high "friendly level".

This is still WIP, so feedback welcome